Create README.md
OS Project 3
Distributed Event Coordination System
Project Report on
Distributed Event Coordination
System
Submitted By-
Atul Gupta
Person # - 5006 0130
atulgupt@buffalo.edu
OS Project 3
Distributed Event Coordination System
Programming Task - Implement a Distributed Event Coordination System.
Language of implementation – C
Data Structure used – I have used graphs and arrays and implemented the graph using the
adjacency list approach.
Example of events relationship represented in the graph -
Event
A
B
C
D
E
After events
B→C→D
C→E
E
Each event is represented using structures.
struct node{
char event;
int nbOfAfterEvents;
struct node *horizontalList[100];
};
For the events list shown as a vertical list on the left is maintained using a array of pointers of
node type.
Each node have an array of pointers of node type (horizontalList). This array will have pointers to
all those nodes (or events) which will occur after this event.
Apart form the array, the node struct have a char type variable which denotes the name of the
event and an integer variable which tells the number of events which occur after this event.
Server implementation details -
Whenever a request is sent by a client it is first sent to a function called parseRequest(). This
function breaks the request into different requests (if there are more than one request).
Eg – insert A->B B->C;query A C;reset;
This request is broken into three different requests.
1) insert A->B B->C
2) query A C
3) reset
The appropriate function is called based on the request type. For insert two functions makeCopy()
and insert() is called. For reset request reset() is called. For query request query() is called.
Insert request – When an insert is requested by the client the following functions will execute.
MakeCopy() - This functions make a copy of the current global graph before making changes.
Insert() - This function breaks the insert request into different requests if there are more than one
inserts are requested.
Eg – insert A->B B->C;
In this example a function insertThisRelatioship() will be called twice, first with A->B and then with
B->C relationship. Then these two relationships will be inserted in the global graph.
OS Project 3
Distributed Event Coordination System
Process of inserting a relation -
Insertion is done in insertRelationship(char before,char after) function. First of all it checks whether
the events are already present or not. Based on their presence in the global graph it execute the
action.
If only before event is present – Function simply adds the after event in the adjacency list of before
event and in the vertical list of the events.
If only after event is present – Function creates a new event in the vertical list i.e. before event
and the after event in its adjacency list.
If none of the events is present – Function creates two new events in the vertical list and adds the
after event in the adjacency list of before event.
If both the events are present – Function checks the existing relationship of the events. If there is
a conflict between the existing relationship and the requested relationship then a “CONFLICT
DETECTED” error is thrown and the entire insert request is not processed.
Undoing a half processed request is done by calling rollBack() function. This function simply
rollbacks the global graph to its earlier image.
Query request – On receiving a query request the two events are sent to findRelationship()
function which returns the appropriate relationship. Before calling this function both the events are
looked up in the vertical list. If either or both of then are not present in the vertical list then
appropriate error message is sent the client.
Finding a relationship – I have used Depth-First-Search algorithm to find the relationship between
two events.
Eg – query A B;
A DFS is run on the node “A”. If the event “B” is found in this process then it is clear that event “A”
happened before “B”.
If B is not found in A's list then DFS is run on node “B”. If event A is found in this process then this
implies “B” happened before “A”.
If none of the above occurs then this means that both the events are concurrent.
Please note that this same function is called by the insert function in case when both the events in
the insert query are already present in the global graph.
Reset request - This is the easiest of all the three requests. We simply delete all the entries from
the global graph.
Eg – insert A->B;query A B;reset;query A B;
Response from server -
INSERT DONE.
A happened before B.
RESET DONE.
Events not found: A, B
Client implementation details -
Client implementation is very simple. The client program simply looks up for the host name
provided (it is taken as “localhost” by default) and the port number (default – 9090) and connects
the client to that server. Giving the host name is not mandatory. As mentioned earlier, if not given
it will consider “localhost” as server.